chore: basic layout
