+++ frontend/src/views/sql-editor/AsidePanel/WorksheetPane/SheetList/common.ts
@@ -3,37 +3,45 @@ import type { SQLEditorTab } from "@/types";
 import type { Worksheet } from "@/types/proto/v1/worksheet_service";
 import { extractWorksheetUID, getHighlightHTMLByRegExp } from "@/utils";
 
-export type TabItem = {
-  type: "TAB";
-  target: SQLEditorTab;
-};
-export type SheetItem = {
-  type: "SHEET";
-  target: Worksheet;
-};
-export type MergedItem = TabItem | SheetItem;
-export const isTabItem = (item: MergedItem): item is TabItem => {
-  return item.type === "TAB";
-};
-export const isSheetItem = (item: MergedItem): item is SheetItem => {
-  return item.type === "SHEET";
+export type ListItemType = "WORKSHEET" | "DRAFT";
+export type ListItem<T extends ListItemType = ListItemType> = {
+  type: T;
+  target: T extends "WORKSHEET"
+    ? Worksheet
+    : T extends "DRAFT"
+      ? SQLEditorTab
+      : never;
 };
 
-export const keyOfItem = (item: MergedItem) => {
-  return isTabItem(item) ? item.target.id : item.target.name;
+export const isTypedListItem = <T extends ListItemType>(
+  type: T,
+  item: ListItem
+): item is ListItem<T> => {
+  return item.type === type;
 };
 
-export const domIDForItem = (item: MergedItem) => {
-  const key = keyOfItem(item);
-  if (isTabItem(item)) {
+export const keyForListItem = (item: ListItem) => {
+  return isTypedListItem("WORKSHEET", item)
+    ? item.target.name
+    : isTypedListItem("DRAFT", item)
+      ? item.target.id
+      : "";
+};
+
+export const domIDForListItem = (item: ListItem) => {
+  const key = keyForListItem(item);
+  if (isTypedListItem("DRAFT", item)) {
     // tab
-    return `bb-sheet-list-tab-${key}`;
+    return `bb-sheet-list-item-${key}`;
   }
-  // sheet
-  return `bb-sheet-list-sheet-${extractWorksheetUID(key)}`;
+  if (isTypedListItem("WORKSHEET", item)) {
+    // sheet
+    return `bb-sheet-list-worksheet-${extractWorksheetUID(key)}`;
+  }
+  return "bb-sheet-list-worksheet-unknown";
 };
 
-export const titleHTML = (item: MergedItem, keyword: string) => {
+export const titleHTML = (item: ListItem, keyword: string) => {
   const kw = keyword.toLowerCase().trim();
 
   const title = item.target.title;
@@ -47,10 +55,4 @@ export const titleHTML = (item: MergedItem, keyword: string) => {
     escape(kw),
     false /* !caseSensitive */
   );
-};
-
-export type DropdownState = {
-  item: MergedItem;
-  x: number;
-  y: number;
 };
