+++ frontend/src/views/sql-editor/AsidePanel/WorksheetPane/SheetListLegacy/index.ts
@@ -1,0 +1,3 @@
+import SheetList from "./SheetList.vue";
+
+export default SheetList;
